requirements.txt
google-generativeai==0.3.2
streamlit==1.30.0
python-dotenv==1.0.0
langchain==0.1.0
langchain_google_genai==0.0.6
faiss-cpu==1.7.4
chromadb==0.4.22
PyPDF2==3.0.1
pdf2image==1.17.0
unstructured==0.12.0